+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 ## 4.43
  - [#365] (https://github.com/cohere-ai/cohere-python/pull/365)
-    - Update embed job create input_dataset param to dataset
+    - Update embed job parameters
 
 ## 4.42
  - [#359] (https://github.com/cohere-ai/cohere-python/pull/359)